Add TRAMP completion support for /ghcs: methods
This simply returns all the names from the codespace list.

I would like to have marked it up with some highlight information like
the name of the associated repository -- I think `codespace-connect`
with completing-read will always be a bit prettier that way -- but this
lets it tie right into the TRAMP interface a bit better.